+++ diffs.js
@@ -11,6 +11,14 @@ function DiffsError( error, stderr )
     this.stderr = stderr;
 }
 
+function uniq( array )
+{
+    return array.sort().filter( function( item, index, array )
+    {
+        return !index || item != array[ index - 1 ];
+    } );
+}
+
 module.exports.fetch = function run( options, document, tempFolder )
 {
     return new Promise( function( resolve, reject )
@@ -46,10 +54,19 @@ module.exports.fetch = function run( options, document, tempFolder )
 
             try
             {
+                var filesWithConflicts = childProcess.execSync( "git diff --name-only --diff-filter=U", { cwd: folder } ).toString().trim().split( '\n' );
                 var relativePath = childProcess.execSync( "git ls-files --full-name " + filePath, { cwd: folder } ).toString().trim();
+                if( filesWithConflicts )
+                {
+                    relativePath = uniq( relativePath.split( '\n' ) )[ 0 ];
+                }
                 options.debug( "Relative path: " + relativePath, options );
-
-                if( relativePath !== "" )
+                if( filesWithConflicts && filesWithConflicts.indexOf( relativePath ) !== -1 )
+                {
+                    options.debug( "Formatting not possible - file currently has merge conflicts", options );
+                    reject( new DiffsError( "Formatting not possible - file currently has merge conflicts" ) );
+                }
+                else if( relativePath !== "" )
                 {
                     var tempFileName = path.join( tempFolder, name );